[crashlogger] Move dso list support to own file.
And move generic utilities into utils.cpp.
