+++ system/core/crashlogger/rules.mk
@@ -11,6 +11,7 @@ MODULE_TYPE := userapp
 MODULE_SRCS += \
     $(LOCAL_DIR)/crashlogger.cpp \
     $(LOCAL_DIR)/backtrace.cpp \
+    $(LOCAL_DIR)/dso-list.cpp \
     $(LOCAL_DIR)/utils.cpp \
 
 MODULE_NAME := crashlogger
